Decouple supported languages object
